Move TypeCounter metatemplate to util/
